Create central implementation of init_generator
