+++ src/emitter/mod.rs
@@ -1,10 +1,11 @@
 pub mod report_var_metadata;
 pub mod rewriting;
-pub mod wizard;
+pub mod module_emitter;
+mod utils;
 #[cfg(test)]
 pub mod tests;
 
-use crate::common::error::WhammError;
+use crate::common::error::{ErrorGen, WhammError};
 use crate::emitter::rewriting::rules::{Arg, LocInfo, WhammProvider};
 use crate::parser::types::{Block, DataType, Expr, Fn, ProbeSpec, Statement, Value};
 
@@ -15,43 +16,54 @@ use wasmparser::Operator;
 // ==== Emitter Trait --> Used By All Emitters! ====
 // =================================================
 
-pub trait ModuleEmitter {
-    fn enter_scope(&mut self) -> Result<(), Box<WhammError>>;
-    fn exit_scope(&mut self) -> Result<(), Box<WhammError>>;
-    fn reset_children(&mut self);
-    fn save_args(&mut self, args: &[Arg]) -> bool;
-    fn emit_args(&mut self) -> Result<bool, Box<WhammError>>;
-    fn define(&mut self, var_name: &str, var_rec: &Option<Value>) -> Result<bool, Box<WhammError>>;
-    fn reset_table_data(&mut self, loc_info: &LocInfo);
-    fn fold_expr(&mut self, expr: &mut Expr) -> bool;
-    fn emit_expr(&mut self, expr: &mut Expr) -> Result<bool, Box<WhammError>>;
-
-    fn emit_fn(&mut self, context_name: &str, f: &Fn) -> Result<bool, Box<WhammError>>;
-    fn emit_formal_param(&mut self, param: &(Expr, DataType)) -> bool;
-    fn emit_global(
+pub trait Emitter {
+    fn emit_body(&mut self, curr_instr_args: &[Arg], body: &mut Block, err: &mut ErrorGen) -> bool;
+    fn emit_stmt(
         &mut self,
-        name: String,
-        ty: DataType,
-        val: &Option<Value>,
-    ) -> Result<bool, Box<WhammError>>;
-    fn remove_orig(&mut self) -> bool;
-    fn emit_orig(&mut self) -> bool;
-    fn emit_if(&mut self) -> bool;
-    fn emit_if_else(&mut self) -> bool;
-    /// Will configure the emitter to emit subsequent expression as the condition of an if or if/else stmt
-    fn emit_condition(&mut self) -> bool;
-    /// Will configure the emitter to emit subsequent statements into the consequent body of an if or if/else stmt
-    fn emit_consequent(&mut self) -> bool;
-    /// Will configure the emitter to emit subsequent statements into the alternate body of an if/else stmt
-    fn emit_alternate(&mut self) -> bool;
-    /// Will configure the emitter to emit subsequent statements in the outer block of some branching logic
-    fn finish_branch(&mut self) -> bool;
-    fn emit_global_stmts(&mut self, stmts: &mut Vec<Statement>) -> Result<bool, Box<WhammError>>;
-    fn emit_body(&mut self, body: &mut Block) -> Result<bool, Box<WhammError>>;
-    fn emit_stmt(&mut self, stmt: &mut Statement) -> Result<bool, Box<WhammError>>;
-
-    fn dump_to_file(&mut self, output_wasm_path: String) -> Result<bool, Box<WhammError>>;
+        curr_instr_args: &[Arg],
+        stmt: &mut Statement,
+        err: &mut ErrorGen,
+    ) -> bool;
+    fn emit_expr(&mut self, expr: &mut Expr, err: &mut ErrorGen) -> bool;
 }
+
+// pub trait ModuleEmitter {
+//     fn enter_scope(&mut self) -> Result<(), Box<WhammError>>;
+//     fn exit_scope(&mut self) -> Result<(), Box<WhammError>>;
+//     fn reset_children(&mut self);
+//     fn save_args(&mut self, args: &[Arg]) -> bool;
+//     fn emit_args(&mut self) -> Result<bool, Box<WhammError>>;
+//     fn define(&mut self, var_name: &str, var_rec: &Option<Value>) -> Result<bool, Box<WhammError>>;
+//     fn reset_table_data(&mut self, loc_info: &LocInfo);
+//     fn fold_expr(&mut self, expr: &mut Expr) -> bool;
+//     fn emit_expr(&mut self, expr: &mut Expr) -> Result<bool, Box<WhammError>>;
+//
+//     fn emit_fn(&mut self, context_name: &str, f: &Fn) -> Result<bool, Box<WhammError>>;
+//     fn emit_formal_param(&mut self, param: &(Expr, DataType)) -> bool;
+//     fn emit_global(
+//         &mut self,
+//         name: String,
+//         ty: DataType,
+//         val: &Option<Value>,
+//     ) -> Result<bool, Box<WhammError>>;
+//     fn remove_orig(&mut self) -> bool;
+//     fn emit_orig(&mut self) -> bool;
+//     fn emit_if(&mut self) -> bool;
+//     fn emit_if_else(&mut self) -> bool;
+//     /// Will configure the emitter to emit subsequent expression as the condition of an if or if/else stmt
+//     fn emit_condition(&mut self) -> bool;
+//     /// Will configure the emitter to emit subsequent statements into the consequent body of an if or if/else stmt
+//     fn emit_consequent(&mut self) -> bool;
+//     /// Will configure the emitter to emit subsequent statements into the alternate body of an if/else stmt
+//     fn emit_alternate(&mut self) -> bool;
+//     /// Will configure the emitter to emit subsequent statements in the outer block of some branching logic
+//     fn finish_branch(&mut self) -> bool;
+//     fn emit_global_stmts(&mut self, stmts: &mut Vec<Statement>) -> Result<bool, Box<WhammError>>;
+//     fn emit_body(&mut self, body: &mut Block) -> Result<bool, Box<WhammError>>;
+//     fn emit_stmt(&mut self, stmt: &mut Statement) -> Result<bool, Box<WhammError>>;
+//
+//     fn dump_to_file(&mut self, output_wasm_path: String) -> Result<bool, Box<WhammError>>;
+// }
 
 pub trait VisitingEmitter {
     fn enter_scope(&mut self) -> Result<(), Box<WhammError>>;
